refactoring : gnl 라이브러리화
